requirements.txt
Django==4.1.5
django_phonenumber_field==7.0.2
django_resized==1.0.2
